pyproject.toml
[tool.poetry]
name = 'blogproject'
version = '0.1.0'
description = ""
authors = ['yan <korotkevich3007@gmail.com>']
readme = 'README.md'

[tool.poetry.dependencies]
python = '^3.12'
django = "^5.1.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
drf-yasg = "^1.21.7"
setuptools = "^74.1.2"
djoser = "^2.2.3"

[build-system]
requires = ['poetry-core']
build-backend = "poetry.core.masonry.api"